edited utils
